LayoutTests/storage/indexeddb/modern/transaction-scheduler-4.html

This test opens a long running read-only transaction to an object store, and then a read-write transaction to the same object store.<br>
The read-write transaction should be blocked while the long running read-only transaction is in progress.<br>
It also queues up a second read-only transaction to that same object store behind the blocked read-write transaction.<br>
Even though that second read-only transaction could start while the first is in progress and the read-write transaction is blocked,<br>
it should *not* start because doing so would risk continuing to block the read-write transaction.<br>
It also queues up a read-only to a different object store behind those first three transactions.<br>
That last read-only transaction *should* start while the read-write transaction is still blocked,<br>
as doing so won't risk blocking the read-write transaction further.<br>
<script>

if (window.testRunner) {
    testRunner.waitUntilDone();
    testRunner.dumpAsText();
}

function done()
{
    alert("Done");
    if (window.testRunner)
        testRunner.notifyDone();
}

var createRequest = window.indexedDB.open("TransactionScheduler4Database");

createRequest.onupgradeneeded = function(event) {
    alert("Upgrade needed: Old version - " + event.oldVersion + " New version - " + event.newVersion);

    var versionTransaction = createRequest.transaction;
    var database = event.target.result;
    var objectStore = database.createObjectStore("OS");
    var request = objectStore.put("bar", "foo");

    request.onerror = function(event) {
        alert("first put FAILED - " + event);
        done();
    }

    objectStore = database.createObjectStore("OtherOS");
    request = objectStore.put("bar", "foo");
    
    request.onerror = function(event) {
        alert("second put FAILED - " + event);
        done();
    }
    
    versionTransaction.onabort = function(event) {
        alert("versionchange transaction aborted");
        done();
    }

    versionTransaction.oncomplete = function(event) {
        alert("versionchange transaction completed");
        continueTest();
        database.close();
    }

    versionTransaction.onerror = function(event) {
        alert("versionchange transaction error'ed - " + event);
        done();
    }
}

var secondDatabaseConnection;
function continueTest()
{
    var longRunningReadRequest = window.indexedDB.open("TransactionScheduler4Database", 1);
    longRunningReadRequest.onsuccess = function(event) {
        alert("Success opening database connection - Starting readonly transaction");
        secondDatabaseConnection = event.target.result;
        
        var transaction = secondDatabaseConnection.transaction("OS", "readonly");
        
        transaction.onerror = function(event) {
            alert("Unexpected transaction error - " + event);
            done();
        }

        transaction.onabort = function(event) {
            alert("Unexpected transaction abort - " + event);
            done();
        }

        transaction.oncomplete = function(event) {
            alert("Read transaction complete - " + event);
        }
        
        firstReadTransactionLoop(transaction, true);
    }
    longRunningReadRequest.onerror = function(event) {
        alert("Long running read request unexpected error - " + event);
        done();
    }
    longRunningReadRequest.onblocked = function(event) {
        alert("Long running read request unexpected blocked - " + event);
        done();
    }
    longRunningReadRequest.onupgradeneeded = function(event) {
        alert("Long running read request unexpected upgradeneeded - " + event);
        done();
    } 
}

var shouldEndReadTransaction = false;

function firstReadTransactionLoop(transaction, isFirstTime)
{
    var objectStore = transaction.objectStore("OS");
    var request = objectStore.get("foo");
    
    request.onsuccess = function(event) {
        if (shouldEndReadTransaction)
            return;
            
        firstReadTransactionLoop(event.target.transaction, false);
        
        // Now that the read transaction is open, the other transactions can be queued up.
        if (isFirstTime)
             startOtherTransactions()
    }

    request.onerror = function(event) {
        alert("Unexpected request error - " + event);
        done();
    }
}

var shouldEndReadTransaction = false;
var loggedWritePutSuccessOnce = false;
function writeTransactionLoop(transaction)
{
    var objectStore = transaction.objectStore("OS");
    var request = objectStore.put("baz", "foo");

    request.onsuccess = function(event) {
        if (!loggedWritePutSuccessOnce) {
            alert("Write transaction first put success");
            loggedWritePutSuccessOnce = true;
        }
        if (shouldEndReadTransaction)
            return;
        
        writeTransactionLoop(transaction);
    }
    
    request.onerror = function(event) {
        alert("Write transaction put unexpected error - " + event);
        done();
    }
}

function startOtherTransactions()
{
    var transaction = secondDatabaseConnection.transaction("OS", "readwrite");

    transaction.onerror = function(event) {
        alert("Write transaction unexpected error - " + event);
        done();
    }

    transaction.onabort = function(event) {
        alert("Write transaction unexpected abort - " + event);
        done();
    }

    transaction.oncomplete = function(event) {
        alert("Write transaction complete - " + event);
    }
    
    writeTransactionLoop(transaction);

    // This read-only transaction should be blocked behind the read-write transaction above, 
    // and should not start until it finishes.
    transaction = secondDatabaseConnection.transaction("OS", "readonly");
    transaction.onerror = function(event) {
        alert("Second read-only transaction unexpected error - " + event);
        done();
    }

    transaction.onabort = function(event) {
        alert("Second read-only transaction unexpected abort - " + event);
        done();
    }

    transaction.oncomplete = function(event) {
        alert("Second read-only transaction complete - " + event);
        done();
    }
    
    var objectStore = transaction.objectStore("OS");
    var request = objectStore.get("foo");

    request.onsuccess = function(event) {
        alert("Second read-only transaction get success - " + event);
    }
    
    request.onerror = function(event) {
        alert("Second read-only transaction put unexpected error - " + event);
        done();
    }

    // This read-only transaction is queued behind the 3 previous transactions, but should not
    // be blocked because its scope doesn't overlap the read-write transaction.
    transaction = secondDatabaseConnection.transaction("OtherOS", "readonly");

    transaction.onerror = function(event) {
        alert("Other objectstore read transaction unexpected error - " + event);
        done();
    }

    transaction.onabort = function(event) {
        alert("Other objectstore read transaction unexpected abort - " + event);
        done();
    }

    transaction.oncomplete = function(event) {
        alert("Other objectstore read transaction complete - " + event);
        shouldEndReadTransaction = true;
    }

    objectStore = transaction.objectStore("OtherOS");
    request = objectStore.get("foo");

    request.onsuccess = function(event) {
        alert("Other objectstore read transaction get success - " + event);
    }
    
    request.onerror = function(event) {
        alert("Other objectstore read transaction get unexpected error - " + event);
        done();
    }
}

</script>
